move the test to the commatrix_test file
